+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next version
 
 - Tasks : add :on_warning_or_alert scope/filter
+- Tasks : fix redirect after create comment
 
 ## 3.9.5